[EAD-15] Refactored PathHelper; extracted subsidiary classes
