+++ core/src/main/java/com/exadel/etoolbox/anydiff/comparison/path/SidedPosition.java
@@ -1,0 +1,39 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License").
+ * You may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.exadel.etoolbox.anydiff.comparison.path;
+
+import com.github.difflib.text.DiffRow;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+/**
+ * Stores a position within a collection of {@link DiffRow} objects and the side of the difference
+ */
+@AllArgsConstructor(access = AccessLevel.PACKAGE)
+@Getter
+class SidedPosition {
+    static final SidedPosition INVALID = new SidedPosition(-1, Side.ANY);
+
+    private int value;
+    private Side side;
+
+    boolean isValid() {
+        return value >= 0;
+    }
+
+    SidedPosition stepBack() {
+        return new SidedPosition(value - 1, side);
+    }
+}
